# swagger.json
{"openapi":"3.1.0","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"paths":{"/api/videos":{"get":{"tags":["video-controller"],"summary":"Get all videos","operationId":"getVideos","responses":{"200":{"description":"Videos found","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/VideoExposureDTO"}}}}}}}},"/api/videos/name/{videoName}":{"get":{"tags":["video-controller"],"summary":"Get videos by name","operationId":"getVideoById","parameters":[{"name":"videoName","in":"path","description":"Name of the video","required":true,"schema":{"type":"string"},"example":"Video.mp4"},{"name":"showAnalytics","in":"query","description":"Indicate whether to show analytics","required":false,"schema":{"type":"boolean","default":false},"example":false},{"name":"showFrames","in":"query","description":"Indicate whether to show frames","required":false,"schema":{"type":"boolean","default":true},"example":true},{"name":"showNotes","in":"query","description":"Indicate whether to show notes","required":false,"schema":{"type":"boolean","default":false},"example":false}],"responses":{"200":{"description":"Videos found","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/VideoExposureDTO"}}}}},"404":{"description":"Videos not found","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/VideoExposureDTO"}}}}}}}},"/api/videos/id/{videoId}":{"get":{"tags":["video-controller"],"summary":"Get video by id","operationId":"getVideoById_1","parameters":[{"name":"videoId","in":"path","description":"ID of the video","required":true,"schema":{"type":"integer","format":"int32"},"example":5},{"name":"showAnalytics","in":"query","description":"Indicate whether to show analytics","required":false,"schema":{"type":"boolean","default":false},"example":false},{"name":"showFrames","in":"query","description":"Indicate whether to show frames","required":false,"schema":{"type":"boolean","default":true},"example":true},{"name":"showNotes","in":"query","description":"Indicate whether to show notes","required":false,"schema":{"type":"boolean","default":false},"example":false}],"responses":{"200":{"description":"Video found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/VideoExposureDTO"}}}},"404":{"description":"Video not found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/VideoExposureDTO"}}}}}}},"/api/tags":{"get":{"tags":["tag-controller"],"summary":"Get all tags","operationId":"getTags","responses":{"200":{"description":"Tags found","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/TagExposureDTO"}}}}}}}},"/api/tags/videoId/{videoId}":{"get":{"tags":["tag-controller"],"summary":"Get all tags on video by video id","operationId":"getTagsByVideoId","parameters":[{"name":"videoId","in":"path","description":"ID of the video","required":true,"schema":{"type":"integer","format":"int32"},"example":5}],"responses":{"200":{"description":"Tags found","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/TagExposureDTO"}}}}},"404":{"description":"Video not found","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/TagExposureDTO"}}}}}}}},"/api/tags/name/{name}":{"get":{"tags":["tag-controller"],"summary":"Get tag by name","operationId":"getTagByName","parameters":[{"name":"name","in":"path","description":"Name of the video","required":true,"schema":{"type":"string"},"example":"Tag1"}],"responses":{"200":{"description":"Tag found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TagExposureDTO"}}}},"404":{"description":"Tag not found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TagExposureDTO"}}}}}}},"/api/tags/id/{tagId}":{"get":{"tags":["tag-controller"],"summary":"Get tag by id","operationId":"getTagById","parameters":[{"name":"tagId","in":"path","description":"ID of the tag","required":true,"schema":{"type":"integer","format":"int32"},"example":5}],"responses":{"200":{"description":"Tag found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TagExposureDTO"}}}},"404":{"description":"Tag not found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TagExposureDTO"}}}}}}},"/api/notes/videoId/{videoId}":{"get":{"tags":["comment-controller"],"summary":"Get comments of a video by video id","operationId":"getNotesByVideoId","parameters":[{"name":"videoId","in":"path","description":"ID of the video","required":true,"schema":{"type":"integer","format":"int32"},"example":5}],"responses":{"200":{"description":"Comments found","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CommentExposureDTO"}}}}},"404":{"description":"Video not found","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CommentExposureDTO"}}}}}}}},"/api/notes/commentId/{commentId}":{"get":{"tags":["comment-controller"],"summary":"Get comment by id","operationId":"getNoteById","parameters":[{"name":"commentId","in":"path","description":"ID of the comment","required":true,"schema":{"type":"integer","format":"int32"},"example":5}],"responses":{"200":{"description":"Comment found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CommentExposureDTO"}}}},"404":{"description":"Comment not found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CommentExposureDTO"}}}}}}},"/api/frames/videoId/{videoId}":{"get":{"tags":["frame-controller"],"summary":"Get all frames of a video by video id","operationId":"getFramesByVideoId","parameters":[{"name":"videoId","in":"path","description":"ID of the video","required":true,"schema":{"type":"integer","format":"int32"},"example":5}],"responses":{"200":{"description":"Frames found","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/FrameExposureDTO"}}}}},"404":{"description":"Video not found","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/FrameExposureDTO"}}}}}}}},"/api/frames/videoId/{videoId}/frameNo/{frameNo}":{"get":{"tags":["frame-controller"],"summary":"Get frame of a video by video id and frame number","operationId":"getFrameByVideoIdAndFrameNumber","parameters":[{"name":"videoId","in":"path","description":"ID of the video","required":true,"schema":{"type":"integer","format":"int32"},"example":5},{"name":"frameNo","in":"path","description":"Number of the frame in the sequence","required":true,"schema":{"type":"integer","format":"int32"},"example":3}],"responses":{"200":{"description":"Frame found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/FrameExposureDTO"}}}},"400":{"description":"Invalid frame number","content":{"*/*":{"schema":{"$ref":"#/components/schemas/FrameExposureDTO"}}}},"404":{"description":"Video not found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/FrameExposureDTO"}}}}}}}},"components":{"schemas":{"CommentExposureDTO":{"type":"object","description":"Output model for a comment on video.","properties":{"id":{"type":"integer","format":"int32","description":"ID of the comment","example":1},"content":{"type":"string","description":"Content of the comment","example":"This is a nice comment!"}}},"FrameExposureDTO":{"type":"object","description":"Output model for a frame of a video.","properties":{"videoId":{"type":"integer","format":"int32","description":"ID of the frame","example":1},"frameNumber":{"type":"integer","format":"int32","description":"Number of the frame in sequence","example":1},"tags":{"type":"array","description":"Tags on the frame","items":{"$ref":"#/components/schemas/TagExposureDTO"}}}},"TagExposureDTO":{"type":"object","description":"Output model for a tag","properties":{"id":{"type":"integer","format":"int32","description":"ID of the tag","example":1},"name":{"type":"string","description":"Name of the tag","example":"Tag"},"value":{"type":"number","format":"double","description":"Value of the tag","example":1.0},"description":{"type":"string","description":"Description of the tag","example":"This is a tag"},"visible":{"type":"boolean","description":"Status of the tag","example":true},"amountUsed":{"type":"integer","format":"int32","description":"Analytics value of how many times the tag has been used on given video","example":2},"totalPoints":{"type":"number","format":"double","description":"Analytics value of total sum of points of given tag on given video","example":37.3}}},"VideoExposureDTO":{"type":"object","description":"Output model of a video","properties":{"id":{"type":"integer","format":"int32","description":"ID of the video","example":1},"name":{"type":"string","description":"Name of the video","example":"Video34.mp4"},"metadata":{"$ref":"#/components/schemas/VideoMetadataExposureDTO","description":"Metadata of the video"},"frames":{"type":"array","description":"Frames of the video","items":{"$ref":"#/components/schemas/FrameExposureDTO"}},"notes":{"type":"array","description":"Comments on the video","items":{"$ref":"#/components/schemas/CommentExposureDTO"}},"uniqueTags":{"type":"integer","format":"int32","description":"Analytics value of how many unique tags are used on the video","example":7},"totalPoints":{"type":"number","format":"double","description":"Analytics value of the total sum of all values of tags used on the video","example":3250.5},"complexity":{"type":"number","format":"double","description":"Analytics value of total calculated complexity of the video","example":3345.23}}},"VideoMetadataExposureDTO":{"type":"object","description":"Output model of metadata of a video","properties":{"totalFrames":{"type":"integer","format":"int32","description":"Total frames of the video","example":350},"frameRate":{"type":"number","format":"double","description":"Framerate of the video","example":24.0},"duration":{"type":"number","format":"double","description":"Duration, in seconds, of the video","example":3.58}}}}}}